+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.3.0 (Unreleased)
 
 * Added edge gateway create/delete functions [#130](https://github.com/vmware/go-vcloud-director/issues/130).
+* Added load balancer service monitor [#196](https://github.com/vmware/go-vcloud-director/pull/196)
 
 ## 2.2.0 (May 15, 2019)